Split GLSL validation tests to a different suite
Many tests in GLSLTest.cpp expected compilation to fail, but in a
fragile way.  Any mistake in the shader could have caused a compilation
failure, and it was up to the author to manually verify that the shader
is fine except for the failure the test was written for.

The new test suite looks at the shader's info log and verifies that the
expected compile error is present.  This change is also in preparation
for porting any compiler unit tests that are written for this purpose.

Bug: angleproject:454531087
Change-Id: I7a3f5ee10a34d43bf94a835e62cb9f70c43d2ca0
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7082029
Reviewed-by: Geoff Lang <geofflang@chromium.org>