Vulkan: Optimize buffer size for vertex attributes
  A previous change related to robust buffer access (CL:6996428),
aimed to fix the rendering issues due to out-of-bounds access on
some platforms by using the entire bufferHelper object size. While
it did resolve the rendering issues on those platforms, it seems to
lead to some WebGL failures on other platforms (e.g., Fuchsia; as
seen in CL:7082305), even if it does not use BindVertexBuffers2().
At least one failed test seemed to be designed for one attribute to
be out-of-bounds so, based on WebGL spec, either an in-bound value
or 0 is used for it.

It is also noteworthy that the platforms showing the robust buffer
access issue (the targets of the original CL) require padding so the
last stride also falls within their range in order not to be seen as
out of bounds. This change will update the size to keep the last
attribute for those platforms in-bound, while attempting to prevent
errors based on the larger size provided.

* Updated bufferSize in syncDirtyEnabledNonStreamingAttrib() in
  VertexArrayVk to bufferVk, with max vertex stride added if the
  platform uses this feature: padBuffersToMaxVertexAttribStride

* Added renderer function to add padding to buffer size if needed:
  padVertexAttribBufferSizeIfNeeded()

* Removed getMaxVertexAttribStride().

Bug: angleproject:448047351
Change-Id: Ib3f97e9f9f87dea5e254b9f292ece200dcc08469
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7099079
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Amirali Abdolrashidi <abdolrashidi@google.com>