[WebGPU] Use actual format ID when getting a shader module.
This CL changes the PipelineKey used in UtilsWgpu's cache to
now store a boolean value that tracks whether the intended
format has alpha bits or not. It makes more sense to use the
actual format ID when generating the wgsl shader for pipelines,
and the intended format was only stored to check whether it had
alpha bits that should be copied over or not.

Bug: angleproject:438268609
Change-Id: I93af73f6933973c9c05838ee5b675b3b7c2710c0
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7095077
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>