Vulkan: Avoid unnecessary loop in FramebufferVk::invalidateImpl
FramebufferVk::invalidateImpl() has some code path does this:
for (size_t colorIndexGL : mState.getEnabledDrawBuffers())
{
   if(invalidateColorBuffers.test(colorIndexGL))
   { ...

If it only invalidates depth or stencil buffer, this loop is completely
a waste. This CL improves this code by bitwise AND
invalidateColorBuffers and mState.getEnabledDrawBuffers() so that we the
loop will early out this the above scenario.

Bug: b/439073246
Change-Id: Ic1be23c49accfe1bb348acfc393864d32f969188
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7106302
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>