Vulkan: Enable OVR_multiview_multisampled_render_to_texture support
This change enables the support for GLES extension
OVR_multiview_multisampled_render_to_texture on vulkan backend.
The extension is enabled when vulkan driver supports both
Multisample Render To Texture (MSRTT) and multiview through
vulkan extensions.

When frontend calls glFramebufferTextureMultisampleMultiviewOVR(),
we pass the numViews and samples to the vulkan backend to call
vkBeginRendering with the proper viewMask and
VkMultisampledRenderToSingleSampledInfoEXT.

Since The callee function Framebuffer::setAttachment() can be called
by other gl entry points that enable only multiview or MSRTT,
but not both, this change adds assert in the backends (GL and D3D) that
support those gl entry points to ensure only one of the multiview and
MSRTT is applied on those backends.

Bug: angleproject:42261787
Change-Id: I0d5e30d6accf8f9c1c5de86c2be179fb2c74b0a3
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7076000
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Yuxin Hu <yuxinhu@google.com>