Vulkan: Use host-image-copy as much as pre-unified-layouts
Prior to VK_KHR_unified_image_layouts, if the driver only exposed a
limited set of layouts to be used with host-image-copy, it was
practically only used during texture initialization.

After VK_KHR_unified_image_layouts, because the driver is required to
expose practically all image layouts as host-copyable, host-image-copy
is used more frequently, including reuploading to textures that have
been used by the GPU, but aren't anymore at the point of upload.

This can be beneficial to some apps, but can hurt CPU-bottlenecked apps.
To be able to choose between the new and old behaviors, a feature is
added.

Bug: angleproject:422982681
Change-Id: I9c21267873bedb9af24349ad1e4e411660d83d4a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7112958
Reviewed-by: Charlie Lao <cclao@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>