Translator: Remove a bunch of redundant compiler tests
Lots of compiler success tests are redundant, because there are either
end2end or deqp tests that not only compile a similar (or more
complicated) shader, but also test that the shader functions correctly.
Those are removed.

If I noticed any hint of a possibly missing functional test, the
compiler tests are retained to be converted to end2end tests as a follow
up.

In some shader-extension test suites, additionally tests that try to
compile ESSL 300 shaders in a GLES 2 context, or tests that take an ESSL
100 shader and prepend `#version 300 es` to it are removed, because they
always failed regardless of what extension is used.

Bug: angleproject:454531087
Change-Id: I492097de06715ac039ab53830865817ae5f3a258
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7113862
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>