Vulkan: Disable preferSubmitAtFBOBoundary
This feature flag was added to improve trace performance when running
back to back in benchmark mode. ARM disabled this flag recently.
Qualcomm also want it disabled in favor of reducing power. The point is
that real app running with vsync will not get benefit of submitting
frame early, but will potentially use more power since now GPU has to
kept power on longer. Also this code path has become too complex
recently due to various tweaks. This CL disables this feature flag
completely. I still kept it around for now in case we need to do
comparison. The expectation is that some traces performance will
regress, but we are going to accept it with the understanding that this
change will benefit real app running in vsync mode.

In future we may need to look into use VkEvent to solve the
bufferSubData kind of situation so that we could still get into fast
path as much as we can.

Bug: b/447633563
Bug: b/422507974
Change-Id: Id6bf0f82038b13780dd8d3beb3ef601952181422
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7107281
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>