IR: Allow dynamic toggling if the IR is built
This change prepares for experimentation when the IR is built by
default.  It also simplifies setting up test expectations in the
meantime by linking the IR usage with a feature flag.

As a side effect, the compiler tests no longer run with the IR.  This is
fixed together with http://anglebug.com/349994211.

Bug: angleproject:349994211
Bug: angleproject:454531087
Change-Id: Ifa6e8bf22c21ca81d98b38cf7c11192c1279e3ab
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7152349
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>