Refactor SupportedSampleSet into a bitfield.
SupportedSampleSet used to be an std::set which was allocated for every
TextureCaps struct for every texture format. This amounted to ~225kb of
std::set allocations per display.

All of this information can be packed into 6 bits since we only use
power-of-two sample counts and a max sample count of 32.

Bug: angleproject:461382864
Change-Id: I67497379f5f4ba0119190cd7e7b2749ab169fa40
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7161916
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>