Disable ResourceMap locking for buffers with no global lock
When the shared context lock is disabled ANGLE is not thread safe so
also disable locking in ResourceMap.

Lower the initial resource allocation size in this case too and count on
re-allocations. Chromium uses fairly few total resources and grows
slowly so the large up-front allocations are costly.

This avoids a ~200kb allocation for each share group.

Bug: angleproject:461382864
Change-Id: I57349dbeac1144038b7f4cc239dad3dd9160236d
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7164155
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>