Vulkan: Expand MSRTSS preference across devices
  The flag preferMSRTSSFlagByDefault affects creating an image with
the MSRTSS flag. If disabled, the created images will not be created
with the MSRTSS flag at first, but will be recreated with this flag
if it is bound to an MSRTT framebuffer object.

The original intent of this flag was to avoid extra memory allocation
on specific drivers. However, currently it is only limited to ARM and
new QCOM drivers. It is possible to enable this flag for all devices
with exceptions.

* Updated preferMSRTSSFlagByDefault to include all devices with the
  multi-sampled-render-to-single-sampled feature support, except the
  older drivers exhibiting the extra memory allocation issue.

Bug: b/461525480
Change-Id: I9929cf134e72f4a1518217bebd9861b0b371f221
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7164225
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Amirali Abdolrashidi <abdolrashidi@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>