WGSL: allow indexing column of uniform matrix
MatCx2s in uniforms need special representation in WGSL,
and they are converted back into regular WGSL matCx2 on use.

When indexing into a uniform matCx2, the WGSL generator was
incorrectly assuming that it was indexing into an array of
matCx2s, and converting the result of the index operation,
instead of performing the conversion on the matCx2 and then
indexing.

Bug: angleproject:461543449
Change-Id: I333c3d525ac8de01196e8cbb1434b41af9d1ad65
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7171072
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>