Add a new BlockLayoutEncoder for packing FP16 uniforms
Create a new BlockLayoutEncoder PackedSPIRVBlockEncoder for
packing default uniforms on vulkan backend tighter. When
convertLowpAndMediumpFloatUniformsTo16BitsThis feature is enabled,
ProgramVk uniform encoder and SPIRV translator encoder are set to use
this new PackedSPIRVBlockEncoder for calculating the offset of each
member within the default uniform block.

The implementations of the PackedSPIRVBlockEncoder are placeholder
functions for now and they work the same as std140BlockEncoder,
so this change should be no op.

Bug: angleproject:405795981
Change-Id: Ia7f3d6594c8666e5610d38c9d2cf0096fa0c2600
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7200889
Commit-Queue: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>