Vulkan: Make ImageHelper::mContentDefined tracks VkImage data
Right now ImageHelper::mContentDefined and mStencilContentDefined is
used to track if ImageHelper has validate data or not. But this is not
consistent. For example, when we stage an update, we do not immediate
set the content valid bit. But when we stage emulated channel clear we
set the bit. In later CL with tile memory I have a need to track whether
VkImage has valid data or not. The decision to reallocate mDeviceMemory
(and data copy) is based on if VkImage has valid data or not. And I will
also add assertion to ensure we never use tile memory when we think
there is valid data. This CL makes it track VkImage have valid data or
not. For this purpose, we renamed mContentDefined to
mVkImageContentDefined and mStencilContentDefined to
mVkImgeStencilContentDefined to make it explicit that it is tracking
VkImage. I have went through all usage of this and ensure that it is
tracking VkImage not ImageHelper. In this CL, stage an update in
ImageHelper will not set the data valid bits. I also removed non-const
version of ImageHelper::getLevelContentDefined() and replaced with
setLevelContentDefined() and clearLevelContentDefined() so that these
variables are well encapsulated and easy to read and search.

Bug: b/447633563
Change-Id: I29164bd2dd603504fc2d21eb1a07daa6a3ed31bd
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7205047
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>