Vulkan: Store mRenderer in WindowSurfaceVk and RenderbufferVk
When ImageHelper changes, we may send notification to owner object which
could be one of WindowSurfaceVk and RenderbufferVk and TextureVk.
Because notification does not have access to mRenderer, there is very
limited things we can do. This CL stores mRenderer in WindowSurfaceVk
and RenderbufferVk so that in later CL I can use it in the
onSubjectStateChange() call.

This CL also made WindowSurfaceVk has its own onSubjectStateChange()
call for later usage. Right now it is doing the exact same thing as
SurfaceVk::onSubjectStateChange(), but that will change in later CL.

Bug: b/447633563
Change-Id: I6fdb365233ff7bdb20672e5260708715f3b58ff1
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7219292
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>