Vulkan: Expand skipped syncVal to include vkCmdEndRenderingKHR
This was exposed in my later CL where I appended a clearTexture call at
end of command buffer submission which runs into this WAW warning. A
test is added to demonstrate this exact usage pattern that produces this
WAW warning. Further investigation indicates that this is a known WAW
and we are already skipping this, except that the skipping was limited
to normal vkCmdEndRenderPass. On my AMD machine it is using dynamic
rendering which uses vkCmdEndRenderingKHR. This CL expands this skip to
also include vkCmdEndRenderingKHR.

Bug: angleproject:400789178
Change-Id: I2b2c6ada1ba324ecc89e8b93075d3736a19ad7fb
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7228645
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>