Translator: Don't use loops to initialize vars without highp
`highp` is practically universally available.  When zero-initializing
variables, this change simply makes the code not use a loop when highp
is not available instead of passing a bool around to so many functions.

The following IR change to do the equivalent transformation will work
similarly.

Bug: angleproject:349994211
Change-Id: Icf2d063735ff315102bf2b3dd975444ba6663b78
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7253954
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>