Vulkan: Remove waitIdleBeforeSwapchainRecreation feature
This feature was originally implemented as a workaround for an
oldSwapchain bug on Android.

The issue occurred when vkCreateSwapchainKHR was called with a valid
oldSwapchain. In this case, the Android framework destroyed the images
in the oldSwapchain. To prevent problems, a workaround was added to
call vkDeviceWaitIdle before recreating the swapchain if oldSwapchain
was valid.

Since Android 12, this bug has been fixed. Therefore, the workaround
is no longer necessary and can be removed.

Bug: angleproject:468061906
Change-Id: Ib13b2ead3cb449d88c6b0a42c4fae0f41b7992af
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7268165
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>