Spanify arguments to ComputeGenericHash()
Then remove the existing function template version of
ComputeGenericHash() in favor of using byte_span_from_ref(), which is
the span-preferred way to access objects as a sequence of bytes.
Otherwise, there may be ambiguity when applying the template to
a container that implicitly converts to span: do we want to write
the span contents, or the meta-data of the container.

Bug: angleproject:436880895
Change-Id: Ia45b5284d2cbf5d0aa4999d09318bde7b81167e1
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7272385
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>