Metal: Increase the maximum number of shader UBOs
While the Desktop GL renderer on M2 MacBook Air can have 16 shader UBOs
at most, the Metal renderer has a hardcoded limit of 12. This
discrepancy causes a compatibility issue.

For instance, when using ANGLE for virglrenderer[1], which provides
OpenGL for virtual machines, VMs see OpenGL ES 3.0 available with the
Desktop GL renderer, but they see it unavailable with the Metal
renderer. This is because virglrenderer reservers one UBO and the VMs
only see 11 UBOs, while OpenGL ES 3.0 requires 12.

Bump the limit hardcoded in the Metal renderer to 16 to match with the
Desktop renderer. It enables the virtualization use case and also avoids
potential problems for other workloads when migrating from the Desktop
GL renderer to the Metal renderer.

[1] https://gitlab.freedesktop.org/virgl/virglrenderer

Bug: angleproject:471257334
Change-Id: I8200d226812f87a8272220cd4122d5c1972a817b
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7319861
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>