Vulkan: Add midRenderPassBlit code path for depthStencil blit
asphalt_9 has a usage pattern that it starts render pass with clear and
then do the blit and then keep rendering. Right now because we picked
blitWithCommands, we end up having to endRenderPass and then blit and
then startRenderPass again. This CL detects midRenderPass blit and bit
by drawing a textured quad, before try with blitWithCommands. This way
we reduces render pass from 3 to 1. A test is added to ensure we don't
regress in future.

Bug: b/470414028
Change-Id: Id35233eb7e4862f7a2419474b648711bfcd9d830
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7489541
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>