Trace/Replay: Orphaned Texture check for ANGLE Capture
OpenGL texture bindings across contexts can result in 'zombie
textures', which are deleted on one context but not on another.
If the original texture ID is used to generate a new texture
of a different type, capture would attempt to create and
initialize the zombie texture for setup which throws GLES
validation errors.

Test: angle_trace_tests --gtest_filter=*life_after
Bug: b/471189378
Change-Id: I0b0187a0ccc39deaac1bd85c0fc5852aba76c5e5
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7309506
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Commit-Queue: Cody Northrop <cnorthrop@google.com>